address warnings
