+++ tests/common/mod.rs
@@ -1,29 +1,14 @@
 //! Utilities and helper functions for integration tests.
+#![allow(dead_code)]
 
 extern crate weld;
 extern crate libc;
-extern crate fnv;
 
-use std::env;
-use std::str;
-use std::slice;
-use std::thread;
-use std::cmp;
-use std::collections::hash_map::Entry;
-
+use weld::*;
 use weld::common::WeldRuntimeErrno;
 
-use weld::WeldConf;
-use weld::WeldValue;
-use weld::WeldError;
-use weld::{weld_value_new, weld_value_data, weld_value_module, weld_value_free};
-use weld::{weld_module_compile, weld_module_run, weld_module_free};
-use weld::{weld_error_new, weld_error_code, weld_error_message, weld_error_free};
-use weld::{weld_conf_new, weld_conf_set, weld_conf_free};
-
-use std::f64::consts::PI;
 use std::ffi::{CStr, CString};
-use libc::{c_char, c_void};
+use self::libc::{c_char, c_void};
 
 /// Compares a and b, and returns true if their difference is less than 0.000...1 (cmp_decimals)
 pub fn approx_equal(a: f64, b: f64, cmp_decimals: u32) -> bool {
@@ -48,7 +33,6 @@ pub fn approx_equal_f32(a: f32, b: f32, cmp_decimals: u32) -> bool {
 
 /// An in memory representation of a Weld vector.
 #[derive(Clone, Debug)]
-#[allow(dead_code)]
 #[repr(C)]
 pub struct WeldVec<T> {
     pub data: *const T,
@@ -81,7 +65,6 @@ impl<T> PartialEq for WeldVec<T> where T: PartialEq + Clone {
 }
 
 #[derive(Clone, Debug, PartialEq)]
-#[allow(dead_code)]
 #[repr(C)]
 pub struct Pair<K, V> {
     pub ele1: K,
